pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "flowdb-ai"
version = "0.1.0"
description = "The AI-Native Database: JSON + Vectors + MCP"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
  { name = "Your Name", email = "your.email@example.com" },
]
keywords = ["database", "ai", "vector", "rag", "mcp", "agent"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "pydantic>=2.0.0",
    "lmdb>=1.4.0",
    "usearch>=2.9.0",
    "numpy",
    "requests",
    "fastmcp",
    "python-dotenv",
    "openai"
]

[project.urls]
Homepage = "https://github.com/yourusername/flowdb"
Issues = "https://github.com/yourusername/flowdb/issues"

[project.scripts]
flowdb = "flowdb.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["flowdb"]